+++ main/src/main/java/ru/practicum/exploreWithMe/model/dto/EventShortDto.java
@@ -8,14 +8,14 @@ import java.time.LocalDateTime;
 
 @Data
 public class EventShortDto {
-    private Integer id;
+    private Long id;
     private String annotation;
     private String title;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
 
-    private Integer category;
+    private Long category;
     private Boolean paid;
     private User initiator;
     private Integer views;
