+++ main/src/main/java/ru/practicum/exploreWithMe/model/dto/EventUserDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class EventUserDto {
-    private Integer id;
+    private Long id;
     private String annotation;
     private String title;
     private String description;
@@ -23,7 +23,7 @@ public class EventUserDto {
     private Integer participantLimit;
     private Boolean requestModeration;
     private State state;
-    private Integer category;
+    private Long category;
     private User initiator;
     private Integer views;
     private Integer confirmedRequests;
