+++ main/src/main/java/ru/practicum/exploreWithMe/model/dto/RequestDto.java
@@ -7,9 +7,9 @@ import java.time.LocalDateTime;
 
 @Data
 public class RequestDto {
-    private Integer id;
+    private Long id;
     private LocalDateTime created;
     private Status status;
-    private Integer requester;
-    private Integer event;
+    private Long requester;
+    private Long event;
 }
