+++ main/src/main/java/ru/practicum/exploreWithMe/repository/CategoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ru.practicum.exploreWithMe.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     @Query("select count(c) from Category c where c.name=:name")
-    Integer getCountByName(String name);
+    Long getCountByName(String name);
 }
